warnings for mingw32 compilation fixed
git-svn-id: https://svn.wxwidgets.org/svn/wx/wxWidgets/trunk@3970 c3d73ce0-8a6f-49c7-b76d-6d57e0e08775

+++ src/msw/listctrl.cpp
@@ -1410,7 +1410,7 @@ bool wxListCtrl::MSWOnNotify(int idCtrl, WXLPARAM lParam, WXLPARAM *result)
     if ( !GetEventHandler()->ProcessEvent(event) )
         return FALSE;
 
-    if (hdr1->code == LVN_GETDISPINFO)
+    if ( (int)hdr1->code == LVN_GETDISPINFO)
     {
         LV_DISPINFO *info = (LV_DISPINFO *)lParam;
         if ( info->item.mask & LVIF_TEXT )
